+++ Sources/MapboxCoreNavigation/RouterDelegate.swift
@@ -132,7 +132,7 @@ public protocol RouterDelegate: AnyObject, UnimplementedLogging {
     /**
      Called when the router fails to receive a new route.
      
-     This method is called after `router(_:willRerouteFrom:)`.
+     This method is called after `router(_:willModify:)`.
      
      - parameter router: The router that has calculated a new route.
      - parameter error: An error raised during the process of obtaining a new route.